Make a AllowInNetdemoPlayback cheat field, move client-specific cheat code to client code
