+++ common/m_cheat.h
@@ -27,39 +27,6 @@
 #include "d_player.h"
 
 //
-// CHEAT SEQUENCE PACKAGE
-//
-
-#ifdef CLIENT_APP
-
-struct cheatseq_t
-{
-	unsigned char* Sequence;
-	unsigned char* Pos;
-	unsigned char DontCheck;
-	unsigned char CurrentArg;
-	unsigned char Args[2];
-	bool (*Handler)(cheatseq_t*);
-};
-
-namespace cheat
-{
-// keycheat handlers
-bool AddKey(cheatseq_t* cheat, unsigned char key, bool* eat);
-
-bool AutoMap(cheatseq_t* cheat);
-bool ChangeLevel(cheatseq_t* cheat);
-bool IdMyPos(cheatseq_t* cheat);
-bool BeholdMenu(cheatseq_t* cheat);
-bool ChangeMusic(cheatseq_t* cheat);
-bool SetGeneric(cheatseq_t* cheat);
-}
-#endif
-
-#define COUNT_CHEATS(l) (sizeof(l) / sizeof(l[0]))
-
-
-//
 // CHEAT TYPES
 //
 enum CheatEnum
